Remove optional fields from configs (#1740)
Closes #1668

Removes optional fields from coin and contract configs. 
Added the `CoinConfigGenerator` to enable easier setup of coin configs
for testing. It holds an internal counter that's used to init the
`tx_id` and `output_index`.

A caveat is that we have to be careful when creating coins/contracts in
unit tests. If we use default values for `tx_id` there could be
overlapping utxos and the node setup will fail. Care also needs to be
applied if tests setup configs from multiple sources.

---------

Co-authored-by: xgreenx <xgreenx9999@gmail.com>