Fetch `ConsensusParameters` from the database (#1809)
Related issue https://github.com/FuelLabs/fuel-core/issues/1753

The change fetches the `ConsensusParameters` from the database and
caches it inside of the `ConsensusParametersProvider`. The
`ConsensusParametersProvider` has a service that runs in the background
and listens for a new block from the block importer. If a new block uses
a new consensus parameters version, the `ConsensusParametersProvider`
will update the internal `latest_consensus_parameters_version` variable
and cache a new `ConsensusParameters`.

### Before requesting review
- [x] I have reviewed the code myself