Use state transition bytecode from the database (#1816)
Related issue https://github.com/FuelLabs/fuel-core/issues/1754

Updated the upgradable executor to fetch the state transition bytecode
from the database when the version doesn't match a native one. This
change enables the WASM executor in the "production" build and requires
a `wasm32-unknown-unknown` target.

The executor has a `VERSION` constant that must be updated when the
behavior of the state transition function is changed. That constant is
used to decide what executor to use(WASM or native).

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself