feat: (re)genesis graceful shutdown (#1821)
closes: #1800 
closes: #1819
closes: #1820 

Resumability:
(Re)genesis can now be paused/resumed. In order to do so the start of
`FuelService` in `cli/run` needed to be split into `starting` and
`running` phases, both interrupt-able by signals.

Progress:
Progress bars are now shown when `stderr` is a `tty`. Otherwise logs
will be printed containing the current group, number of groups, table
name and ETA to completion.

Bars:

![image](https://github.com/FuelLabs/fuel-core/assets/37515857/b3d9467d-85a9-42b0-aa8b-7118ba612ae2)

Logs:

![image](https://github.com/FuelLabs/fuel-core/assets/37515857/d985dd82-d87e-4558-8ca7-935ef8404dc0)


Empty tables:
@MujkicA found an edge case while e2e testing. Fixed here by allowing
tables to not be present in snapshot when using parquet.

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: Hannes Karppila <hannes.karppila@gmail.com>