Support upgradability of the consensus parameters and state transition bytecode in genesis (#1826)
The changes make the state transition bytecode part of the
`ChainConfig`. It guarantees the state transition's availability for the
network's first blocks.

The change has many minor improvements in different areas related to the
state transition bytecode:
- The state transition bytecode lies in its own
file(`state_transition_bytecode.wasm`) along with the chain config file.
The `ChainConfig` loads it automatically when `ChainConfig::load` is
called and pushes it back when `ChainConfig::write` is called.
- The `fuel-core` release bundle also contains the
`fuel-core-wasm-executor.wasm` file of the corresponding executor
version.
- The regenesis process now considers the last block produced by the
previous network. When we create a (re)genesis block of a new network,
it has the `height = last_block_of_old_netowkr + 1`. It continues the
old network and doesn't overlap blocks(before, we had `old_block.height
== new_genesis_block.hegiht`).
- Along with the new block height, the regenesis process also increases
the state transition bytecode and consensus parameters versions. It
guarantees that a new network doesn't use values from the previous
network and allows us not to migrate `StateTransitionBytecodeVersions`
and `ConsensusParametersVersions` tables.
- Added a new CLI argument, `native-executor-version,` that allows
overriding of the default version of the native executor. It can be
useful for side rollups that have their own history of executor
upgrades.
- Replaced:
  
  ```rust
           let file = std::fs::File::open(path)?;
           let mut snapshot: Self = serde_json::from_reader(&file)?;
  ```
  
  with a:
  
  ```rust
           let mut json = String::new();
           std::fs::File::open(&path)
.with_context(|| format!("Could not open snapshot file: {path:?}"))?
               .read_to_string(&mut json)?;
let mut snapshot: Self = serde_json::from_str(json.as_str())?;
  ```
  because it is faster in 100 times for big JSON files.
- Updated all tests to use `Config::local_node_*` instead of working
with the `SnapshotReader` directly. It is the preparation of the tests
for the futures bumps of the `Executor::VERSION`. When we increase the
version, all tests continue to use
`GenesisBlock.state_transition_btyecode = 0` while the version is
different, which forces the usage of the WASM executor, while for tests,
we still prefer to test native execution. The `Config::local_node_*`
handles it and forces the executor to use the native version.
- Reworked the `build.rs` file of the upgradable executor. The script
now has a cache used between builds to avoid recompilation of the WASM
bytecode. Also, fixed the issue with outdated WASM bytecode. The script
reacts on any modifications of the `fuel-core-wasm-executor` and forces
recompilation(it is why we need cache), so WASM bytecode always is
actual now.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Hannes Karppila <hannes.karppila@gmail.com>